+++ src/domain/app/AppInfoDropdown.tsx
@@ -1,6 +1,5 @@
-import { useCallback, useState } from "react";
+import { useCallback, useRef, useState } from "react";
 import { Dropdown } from "react-bootstrap";
-import ReactDOM from "react-dom";
 import { useTranslation } from "react-i18next";
 
 import OutboundLink from "../../common/a11y/OutboundLink";
@@ -8,9 +7,15 @@ import SMIcon from "../../common/components/SMIcon";
 import AppAboutModal from "./AppAboutModal";
 import AppFeedbackModal from "./AppFeedbackModal";
 
+// FIXME: When a modal is open, an escape key press will close the modal and
+//        the dropdown.
 function AppInfoDropdown() {
+  const dropdownRef = useRef<HTMLDivElement | null>(null);
   const { t } = useTranslation();
   const [modal, setModal] = useState<"about" | "feedback" | null>(null);
+  // Manage dropdown state "manually" to force it to remain open when a modal
+  // is opened through it, or when the user tabs onwards from the toggle element.
+  const [isDropdownOpen, setIsDropdownOpen] = useState<boolean>(false);
 
   const handleOnClose = useCallback(() => {
     setModal(null);
@@ -18,7 +23,36 @@ function AppInfoDropdown() {
 
   return (
     <>
-      <Dropdown className="app-info-dropdown" role="contentinfo">
+      <Dropdown
+        ref={dropdownRef}
+        className="app-info-dropdown"
+        role="contentinfo"
+        onToggle={(isOpen, event, { source }) => {
+          // Always open the menu
+          if (isOpen) {
+            return setIsDropdownOpen(isOpen);
+          }
+
+          const isRootClose = source === "rootClose";
+
+          // Only close with a rootClose event when a modal is not open.
+          // By doing this we ensure that the dropdown is not closed when:
+          //
+          // 1) The user tabs from the menu toggle. The default behavior is
+          //    that a user can't access the content of the menu with a tab
+          //    but accessibility auditors often recommend that it is made
+          //    possible
+          //
+          // 2) A modal is opened. When a modal is opened, focus should be moved
+          //    to it. When a modal is closed, the focus should be moved back to
+          //    the element that was used to toggle open the modal. If the menu
+          //    is closed, this can't happen.
+          if (isRootClose && modal === null) {
+            setIsDropdownOpen(isOpen);
+          }
+        }}
+        show={isDropdownOpen}
+      >
         <Dropdown.Toggle>
           <SMIcon icon="info" aria-label={t("APP.ABOUT")} />
         </Dropdown.Toggle>
@@ -45,14 +79,8 @@ function AppInfoDropdown() {
           </Dropdown.Item>
         </Dropdown.Menu>
       </Dropdown>
-      {ReactDOM.createPortal(
-        <>
-          {modal === "about" && <AppAboutModal onClose={handleOnClose} />}
-          {modal === "feedback" && <AppFeedbackModal onClose={handleOnClose} />}
-        </>,
-        // @ts-ignore
-        document.getElementById("modals")
-      )}
+      <AppAboutModal show={modal === "about"} onClose={handleOnClose} />
+      <AppFeedbackModal show={modal === "feedback"} onClose={handleOnClose} />
     </>
   );
 }