+++ src/domain/home/HomeContainer.tsx
@@ -1,24 +1,17 @@
 import className from "classnames";
 import { useCallback, useRef, ReactNode } from "react";
-import { useTranslation } from "react-i18next";
 import { Map as RLMap } from "react-leaflet";
-import { useSelector } from "react-redux";
 import { Switch, useRouteMatch, Route } from "react-router-dom";
 
-import Page from "../../common/a11y/Page";
 import useIsMobile from "../../common/hooks/useIsMobile";
-import useLanguage from "../../common/hooks/useLanguage";
 import ApplicationHeader from "../app/AppHeader";
 import AppInfoDropdown from "../app/AppInfoDropdown";
-import { AppState } from "../app/appConstants";
 import routerPaths from "../app/appRoutes";
 import useIsUnitBrowserSearchView from "../app/useIsUnitBrowserSearchView";
 import Map from "../map/Map";
 import UnitBrowser from "../unit/browser/UnitBrowser";
 import UnitDetails from "../unit/details/UnitDetails";
-import { getUnitById } from "../unit/state/selectors";
-import { Unit } from "../unit/unitConstants";
-import { getAttr, getUnitPosition } from "../unit/unitHelpers";
+import { getUnitPosition } from "../unit/unitHelpers";
 import useFetchInitialData from "./useFetchInitialData";
 
 function useIsUnitDetailsSearchView() {
@@ -30,69 +23,12 @@ function useIsUnitDetailsSearchView() {
   return Boolean(match && match.isExact);
 }
 
-function useHomeMeta() {
-  const { t } = useTranslation();
-  const activeLanguage = useLanguage();
-  const unitDetailsMatch = useRouteMatch<Params>(routerPaths.unitDetails);
-
-  const selectedUnit = useSelector<AppState, Unit>((state) =>
-    getUnitById(state, {
-      id: unitDetailsMatch?.params.unitId,
-    })
-  );
-
-  const getTitle = () => {
-    const title = [];
-
-    if (selectedUnit) {
-      title.push(`${getAttr(selectedUnit.name, activeLanguage) || ""} | `);
-    }
-
-    title.push(t("APP.NAME"));
-
-    return title.join("");
-  };
-
-  const getDescription = () => {
-    if (selectedUnit) {
-      const description = selectedUnit.description;
-
-      if (description) {
-        return getAttr(description, activeLanguage);
-      }
-    }
-
-    return t("APP.DESCRIPTION");
-  };
-
-  const getImage = () => {
-    if (selectedUnit) {
-      return selectedUnit.picture_url;
-    }
-  };
-
-  return {
-    title: getTitle(),
-    description: getDescription(),
-    image: getImage(),
-  };
-}
-
 type MapLayoutProps = {
   content: ReactNode;
   map: ReactNode;
-  title: string;
-  description: string | null;
-  image?: string | null;
 };
 
-function MapLayout({
-  content,
-  map,
-  title,
-  description,
-  image,
-}: MapLayoutProps) {
+function MapLayout({ content, map }: MapLayoutProps) {
   const isUnitSearchOpen = useIsUnitBrowserSearchView();
   const isUnitDetailsOpen = useIsUnitDetailsSearchView();
 
@@ -108,14 +44,7 @@ function MapLayout({
         })}
       >
         <ApplicationHeader />
-        <Page
-          title={title}
-          description={description}
-          image={image}
-          className="map-foreground-content"
-        >
-          {content}
-        </Page>
+        <div className="map-foreground-content">{content}</div>
       </div>
       <AppInfoDropdown />
       <div className="map-container">{map}</div>
@@ -123,15 +52,10 @@ function MapLayout({
   );
 }
 
-type Params = {
-  unitId: string;
-};
-
 function HomeContainer() {
   const mapRef = useRef<RLMap | null>(null);
   const leafletElementRef = useRef<L.Map | null>(null);
   const isMobile = useIsMobile();
-  const { title, description, image } = useHomeMeta();
 
   const handleOnViewChange = useCallback((coordinates) => {
     leafletElementRef.current?.setView(coordinates);
@@ -185,9 +109,6 @@ function HomeContainer() {
 
   return (
     <MapLayout
-      title={title}
-      description={description}
-      image={image}
       content={
         <Switch>
           <Route